Changed world object system
World objects are now based on layers. This means they only need to be drawn once and can then be moved at will without a significant performance impact (as opposed to redrawing everything slightly more to the right).

The system that joins the different layers together should probably be more optimized (especially in situations where one object clearly overwrites pixels below it - if a world object is completely opaque, there is no reason to draw the things below it since they will be covered anyway)

Changed most functions and values from u32 to usize to prevent having to 'as usize' all the values before indexing a vec.